Upgrading to 2020 Setup (#23)
* Initial changes for v2 setup

* Temporary upload version

* Enabling GPU Robomaker

* Updating structure and docker

* Adding default files

* Removing link to Deepracer Workshops

* Removed last external module

* Removing build steps and local dockerfiles

* Remove 'logs proxy'

* Remove automatic upload of logs

* Log-analysis changes

* Updated scripts for v2

* Bugfixing in clean VM

* Bugfixes

* File name changes

* Moving activate.sh

* Bugfix

* Update README.md

* Fixing activation script

* Fix names

* Bugfixing

* Setting CPU as standard for Robomaker

+++ bin/init.sh
@@ -7,8 +7,12 @@ function ctrl_c() {
         exit 1
 }
 
-while getopts ":m:c:" opt; do
+OPT_ARCH="gpu"
+
+while getopts ":m:c:a:" opt; do
 case $opt in
+a) OPT_ARCH="$OPTARG"
+;;
 m) OPT_MOUNT="$OPTARG"
 ;; 
 c) OPT_CLOUD="$OPTARG"
@@ -19,11 +23,14 @@ exit 1
 esac
 done
 
-GPUS=$(docker run --rm --gpus all nvidia/cuda:10.2-base nvidia-smi "-L" | awk  '/GPU .:/' | wc -l)
-if [ $? -ne 0 ] || [ $GPUS -eq 0 ]
+if [[ "${OPT_ARCH}" == "gpu" ]]
 then
-	echo "No GPU detected in docker. Please check setup".
-	exit 1
+    GPUS=$(docker run --rm --gpus all nvidia/cuda:10.2-base nvidia-smi "-L" | awk  '/GPU .:/' | wc -l)
+    if [ $? -ne 0 ] || [ $GPUS -eq 0 ]
+    then
+        echo "No GPU detected in docker. Please check setup".
+        exit 1
+    fi
 fi
 
 INSTALL_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )/.." >/dev/null 2>&1 && pwd )"
@@ -45,32 +52,12 @@ mkdir -p $INSTALL_DIR/docker/volumes
 mkdir -p $(eval echo "~${USER}")/.aws
 ln -sf $(eval echo "~${USER}")/.aws  $INSTALL_DIR/docker/volumes/
 
-# grab local training deepracer repo from crr0004 and log analysis repo from vreadcentric
-# Now as submodules!
-# git clone --recurse-submodules https://github.com/crr0004/deepracer.git
-# git clone https://github.com/breadcentric/aws-deepracer-workshops.git && cd aws-deepracer-workshops && git checkout enhance-log-analysis && cd ..
-git submodule update --init --recursive
-
-ln -sf $INSTALL_DIR/aws-deepracer-workshops/log-analysis  $INSTALL_DIR/docker/volumes/log-analysis
-cp $INSTALL_DIR/deepracer/simulation/aws-robomaker-sample-application-deepracer/simulation_ws/src/deepracer_simulation/routes/* docker/volumes/log-analysis/tracks/
-
 # copy rewardfunctions
-mkdir -p $INSTALL_DIR/custom_files $INSTALL_DIR/analysis
-cp $INSTALL_DIR/deepracer/custom_files/* $INSTALL_DIR/custom_files/
+mkdir -p $INSTALL_DIR/custom_files $INSTALL_DIR/logs $INSTALL_DIR/analysis
 cp $INSTALL_DIR/defaults/hyperparameters.json $INSTALL_DIR/custom_files/
+cp $INSTALL_DIR/defaults/model_metadata.json $INSTALL_DIR/custom_files/
+cp $INSTALL_DIR/defaults/reward_function.py $INSTALL_DIR/custom_files/
 
-# setup symlink to rl-coach config file
-ln -f $INSTALL_DIR/defaults/rl_deepracer_coach_robomaker.py $INSTALL_DIR/deepracer/rl_coach/rl_deepracer_coach_robomaker.py 
-
-# patching files in submodules that don't entirely fit our needs
-cd $INSTALL_DIR/deepracer/
-patch simulation/aws-robomaker-sample-application-deepracer/simulation_ws/src/sagemaker_rl_agent/markov/environments/deepracer_racetrack_env.py < ../defaults/deepracer_racetrack_env.py.patch 
-patch robomaker.env < ../defaults/robomaker.env.patch
-patch rl_coach/env.sh < ../defaults/rl_coach_env.sh.patch
-cd ..
-
-# replace the contents of the rl_deepracer_coach_robomaker.py file with the gpu specific version (this is also where you can edit the hyperparameters)
-# TODO this file should be genrated from a gui before running training
 cp $INSTALL_DIR/defaults/template-run.env $INSTALL_DIR/current-run.env
 if [[ -n "$OPT_CLOUD" ]];
 then
@@ -88,13 +75,10 @@ do
 done
 
 # Download docker images. Change to build statements if locally built images are desired.
-# docker build ${args} -f ./docker/dockerfiles/rl_coach/Dockerfile -t larsll/deepracer-rlcoach ./
-# docker build ./docker/dockerfiles/deepracer_robomaker/ -t larsll/deepracer-robomaker
-# docker build ./docker/dockerfiles/log-analysis/ -t larsll/deepracer-loganalysis
-docker pull larsll/deepracer-rlcoach
-docker pull larsll/deepracer-robomaker
-# docker pull larsll/deepracer-loganalysis
-docker pull crr0004/sagemaker-rl-tensorflow:nvidia
+docker pull larsll/deepracer-rlcoach:v2
+docker pull awsdeepracercommunity/deepracer-robomaker:cpu
+docker pull awsdeepracercommunity/deepracer-sagemaker:$OPT_ARCH
+docker pull larsll/deepracer-loganalysis:v2-cpu
 
 # create the network sagemaker-local if it doesn't exit
 SAGEMAKER_NW='sagemaker-local'
@@ -105,7 +89,7 @@ then
 fi
 
 # ensure our variables are set on startup
-echo "source $INSTALL_DIR/activate.sh" >> $HOME/.profile
+echo "source $INSTALL_DIR/bin/activate.sh" >> $HOME/.profile
 
 # mark as done
 date | tee $INSTALL_DIR/DONE
